# frontend/.openapi.yaml
components:
  schemas:
    ErrorDetail:
      additionalProperties: false
      properties:
        location:
          description: Where the error occurred, e.g. 'body.items[3].tags' or 'path.thing-id'
          type: string
        message:
          description: Error message text
          type: string
        value:
          description: The value at the given location
      type: object
    ErrorModel:
      additionalProperties: false
      properties:
        $schema:
          description: A URL to the JSON Schema for this object.
          examples:
            - /api/schemas/ErrorModel.json
          format: uri
          readOnly: true
          type: string
        detail:
          description: A human-readable explanation specific to this occurrence of the problem.
          examples:
            - Property foo is required but is missing.
          type: string
        errors:
          description: Optional list of individual error details
          items:
            $ref: "#/components/schemas/ErrorDetail"
          type:
            - array
            - "null"
        instance:
          description: A URI reference that identifies the specific occurrence of the problem.
          examples:
            - https://example.com/error-log/abc123
          format: uri
          type: string
        status:
          description: HTTP status code
          examples:
            - 400
          format: int64
          type: integer
        title:
          description: A short, human-readable summary of the problem type. This value should not change between occurrences of the error.
          examples:
            - Bad Request
          type: string
        type:
          default: about:blank
          description: A URI reference to human-readable documentation for the error.
          examples:
            - https://example.com/errors/example
          format: uri
          type: string
      type: object
    GreetingOutputBody:
      additionalProperties: false
      properties:
        $schema:
          description: A URL to the JSON Schema for this object.
          examples:
            - /api/schemas/GreetingOutputBody.json
          format: uri
          readOnly: true
          type: string
        msg:
          description: The response of a greeting
          examples:
            - Hello John Doe!
          type: string
      required:
        - msg
      type: object
    LoginRequest:
      additionalProperties: false
      properties:
        $schema:
          description: A URL to the JSON Schema for this object.
          examples:
            - /api/schemas/LoginRequest.json
          format: uri
          readOnly: true
          type: string
        Token:
          type: string
      required:
        - Token
      type: object
    LoginResponseBody:
      additionalProperties: false
      properties:
        $schema:
          description: A URL to the JSON Schema for this object.
          examples:
            - /api/schemas/LoginResponseBody.json
          format: uri
          readOnly: true
          type: string
        Success:
          type: boolean
      required:
        - Success
      type: object
  securitySchemes:
    token:
      in: cookie
      name: session
      type: apiKey
info:
  title: Greetings API
  version: 1.0.0
openapi: 3.1.0
paths:
  /greeting/{user}:
    get:
      description: Greet a person by name
      operationId: get-greeting
      parameters:
        - example: ""
          in: path
          name: user
          required: true
          schema:
            examples:
              - ""
            maxLength: 15
            type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GreetingOutputBody"
          description: OK
        default:
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ErrorModel"
          description: Error
      security:
        - token:
            - greeting
      summary: Get a greeting
  /login:
    post:
      description: By loggin in a cookie will be assinged required for auth
      operationId: login
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponseBody"
          description: OK
          headers:
            Set-Cookie:
              schema:
                type: string
        default:
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ErrorModel"
          description: Error
      security:
        - token:
            - NO_AUTH
      summary: Log in
servers:
  - url: /api